Split Schlomka2008 and DualEnergy NegLogLikelihoods into two classes, with most of the code in the mother class
